Restructuring the info package
