+++ app/src/main/java/wishcantw/vocabulazy/activities/mainmenu/activity/MainMenuActivity.java
@@ -15,10 +15,10 @@ import wishcantw.vocabulazy.activities.ParentActivity;
 import wishcantw.vocabulazy.database.AppPreference;
 import wishcantw.vocabulazy.activities.exam.activity.ExamActivity;
 import wishcantw.vocabulazy.activities.mainmenu.fragment.MainMenuFragment;
-import wishcantw.vocabulazy.activities.mainmenu.info.ReportPageFragment;
-import wishcantw.vocabulazy.activities.mainmenu.info.TNCFragment;
-import wishcantw.vocabulazy.activities.mainmenu.info.UserGuideFragment;
-import wishcantw.vocabulazy.activities.mainmenu.info.WishCanIntroFragment;
+import wishcantw.vocabulazy.activities.mainmenu.info.fragment.ReportPageFragment;
+import wishcantw.vocabulazy.activities.mainmenu.info.fragment.TNCFragment;
+import wishcantw.vocabulazy.activities.mainmenu.info.fragment.UserGuideFragment;
+import wishcantw.vocabulazy.activities.mainmenu.info.fragment.WishCanIntroFragment;
 import wishcantw.vocabulazy.activities.mainmenu.model.MainMenuModel;
 import wishcantw.vocabulazy.activities.mainmenu.note.fragment.NoteCreateDialogFragment;
 import wishcantw.vocabulazy.activities.mainmenu.note.fragment.NoteDeleteDialogFragment;
