+++ app/src/main/java/wishcantw/vocabulazy/activities/mainmenu/info/fragment/UserGuideFragment.java
@@ -1,4 +1,4 @@
-package wishcantw.vocabulazy.activities.mainmenu.info;
+package wishcantw.vocabulazy.activities.mainmenu.info.fragment;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -14,14 +14,14 @@ import wishcantw.vocabulazy.analytics.ga.GABaseFragment;
  * Created by allencheng07 on 2016/9/21.
  */
 
-public class TNCFragment extends GABaseFragment {
+public class UserGuideFragment extends GABaseFragment {
 
-    public static final String TAG = "TNCFragment";
+    public static final String TAG = "UserGuideFragment";
 
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_tnc, container, false);
+        View view = inflater.inflate(R.layout.fragment_user_guide, container, false);
         findViews(view);
         return view;
     }
@@ -32,6 +32,6 @@ public class TNCFragment extends GABaseFragment {
 
     @Override
     protected String getGALabel() {
-        return Analytics.ScreenName.TNC;
+        return Analytics.ScreenName.USER_GUIDE;
     }
 }
