+++ app/src/main/java/wishcantw/vocabulazy/activities/mainmenu/info/fragment/WishCanIntroFragment.java
@@ -1,4 +1,4 @@
-package wishcantw.vocabulazy.activities.mainmenu.info;
+package wishcantw.vocabulazy.activities.mainmenu.info.fragment;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -9,6 +9,7 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import wishcantw.vocabulazy.R;
+import wishcantw.vocabulazy.activities.mainmenu.info.adapter.MemberInfoAdapter;
 import wishcantw.vocabulazy.analytics.Analytics;
 import wishcantw.vocabulazy.analytics.ga.GABaseFragment;
