+++ app/src/main/res/layout/view_info.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<wishcantw.vocabulazy.activities.mainmenu.info.InfoView xmlns:android="http://schemas.android.com/apk/res/android"
+<wishcantw.vocabulazy.activities.mainmenu.info.view.InfoView xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
@@ -91,4 +91,4 @@
         android:layout_height="5dp"
         android:background="@color/white_2"/>
 
-</wishcantw.vocabulazy.activities.mainmenu.info.InfoView>
+</wishcantw.vocabulazy.activities.mainmenu.info.view.InfoView>
